+++ CHANGELOG.md
@@ -1,2 +1,7 @@
+## Unreleased
+- Adds [`Objectware.isPlainObject`][isPlainObject].
+
+[isPlainObject]: https://github.com/moll/js-objectware/blob/master/doc/API.md#Objectware.isPlainObject
+
 ## 1.0.0 (Jan 13, 2015)
 - Perfectly objective.